refactor(font): extract FontSettings to own file
Move FontSettings data class out of
FontPresetApplicator.kt so Kover/Codecov excludes
target only the untestable IDE-glue object, not
the fully-tested domain logic. Add Codecov ignores
for FontPresetApplicator, FontPresetPanel, and
FontPreviewComponent (IDE singletons / Swing UI).
