+++ codecov.yml
@@ -15,6 +15,9 @@ coverage:
 
 ignore:
   - "src/main/kotlin/dev/ayuislands/licensing/**"
+  - "src/main/kotlin/dev/ayuislands/font/FontPresetApplicator.kt"
+  - "src/main/kotlin/dev/ayuislands/settings/FontPresetPanel.kt"
+  - "src/main/kotlin/dev/ayuislands/settings/FontPreviewComponent.kt"
 
 comment:
   layout: "condensed_header, condensed_files, condensed_footer"
